pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pyAPisolation"
version = "0.7.1"
dependencies = [
    "numpy>=1.21.0, <1.24",
    "h5py>=2.10",
    "pyabf",
    "matplotlib",
    "pandas",
    "scipy",
    "anndata",
    "scikit-learn",
    "joblib",
    "allensdk",
    "plotly",
    "pyyaml",
    "beautifulsoup4"
]


[tool.setuptools.packages.find]
where = ["."]  # list of folders that contain the packages (["."] by default)
include = ["*"]  # package names should match these glob patterns (["*"] by default)
exclude = ["tests"]  # exclude packages matching these glob patterns (empty by default)

[project.scripts]
spike_finder = "pyAPisolation.bin:run_spike_finder.main"
pyAPisolation_setup = "pyAPisolation.setup_install:install_ipfx"
[tool.setuptools.package-data]
"*" = ["*.pzfx", "*.ui"]

[project.optional-dependencies]
gui = ["PySide2", "pyqtgraph"]
web = ["flask"]
server = ["gunicorn", "flask-cors", "flask"]
dev = ["torch"]
full = [
    "PySide2", 
    "pyqtgraph", 
    "flask",
    "gunicorn",
    "flask-cors"
]